[common/desktop] Estimates video file sizes using video length and ideal frame size * 30 per second
In the process refactored every search result class descendant from FileSearchResult, for some reason we had `double` as the type for file size, fixed to long.

+++ common/src/main/java/com/frostwire/search/FileSearchResult.java
@@ -26,5 +26,5 @@ public interface FileSearchResult extends SearchResult {
 
     String getFilename();
 
-    double getSize();
+    long getSize();
 }
