+++ common/src/main/java/com/frostwire/search/limetorrents/LimeTorrentsSearchResult.java
@@ -1,6 +1,6 @@
 /*
  * Created by Angel Leon (@gubatron), Alden Torres (aldenml), alejandroarturom
- * Copyright (c) 2011-2020, FrostWire(R). All rights reserved.
+ * Copyright (c) 2011-2023, FrostWire(R). All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -34,7 +34,7 @@ public final class LimeTorrentsSearchResult extends AbstractTorrentSearchResult 
     private final String detailsUrl;
     private final String torrentUrl;
     private final String infoHash;
-    private final double size;
+    private final long size;
     private final long creationTime;
     private final int seeds;
 
@@ -57,7 +57,7 @@ public final class LimeTorrentsSearchResult extends AbstractTorrentSearchResult 
     }
 
     @Override
-    public double getSize() {
+    public long getSize() {
         return size;
     }
