+++ common/src/main/java/com/frostwire/search/tpb/TPBSearchResult.java
@@ -1,12 +1,12 @@
 /*
  * Created by Angel Leon (@gubatron), Alden Torres (aldenml)
- * Copyright (c) 2011-2017, FrostWire(R). All rights reserved.
+ * Copyright (c) 2011-2023, FrostWire(R). All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -56,7 +56,7 @@ public class TPBSearchResult extends AbstractTorrentSearchResult {
     private final String detailsUrl;
     private final String torrentUrl;
     private final String infoHash;
-    private final double size;
+    private final long size;
     private final long creationTime;
     private final int seeds;
 
@@ -90,7 +90,7 @@ public class TPBSearchResult extends AbstractTorrentSearchResult {
     }
 
     @Override
-    public double getSize() {
+    public long getSize() {
         return size;
     }
 
@@ -129,7 +129,7 @@ public class TPBSearchResult extends AbstractTorrentSearchResult {
         return displayName;
     }
 
-    protected double parseSize(String group) {
+    protected long parseSize(String group) {
         String[] size = group.split("&nbsp;");
         String amount = size[0].trim();
         String unit = UNIT_MAPPER.get(size[1].trim());
